Removed core.py and replaced it with modelgrid.py and utils.py to employ better named modules and isolate classes/
